homework6: fixes

+++ homework6MethodOverloading/CalculateIncomeTaxBasedOnSalary.java
@@ -2,7 +2,7 @@ package homework6MethodOverloading;
 
 public class CalculateIncomeTaxBasedOnSalary {
     public static void main(String[] args) {
-        int grossSalary = 500;
+        double grossSalary = 500;
         int customTaxRate = 15;
         boolean isEmployeeITWorker = true;
 
@@ -13,17 +13,17 @@ public class CalculateIncomeTaxBasedOnSalary {
         System.out.println("Tax of gross salary with default(20%) or special(10%) tax rate: " + calculateIncomeTax(grossSalary, isEmployeeITWorker));
     }
 
-    static double calculateIncomeTax(int grossSalary, int taxRateInPercent) {
-        double taxRate = (taxRateInPercent / 100.0);
+    static double calculateIncomeTax(double grossSalary, double taxRateInPercent) {
+        double taxRate = (taxRateInPercent / 100);
         return grossSalary * taxRate;
     }
 
-    static double calculateIncomeTax(int grossSalary) {
+    static double calculateIncomeTax(double grossSalary) {
         // set the default tax rate 20%
         return calculateIncomeTax(grossSalary, 20);
     }
 
-    static double calculateIncomeTax(int grossSalary, boolean isITSectorEmployee) {
+    static double calculateIncomeTax(double grossSalary, boolean isITSectorEmployee) {
         // set the special tax rate 10% if employee is IT worker
         return isITSectorEmployee?
                 calculateIncomeTax(grossSalary, 10):
